requirements.txt
streamlit==1.28.2
transformers==4.46.2
torch==2.5.0
Pillow==10.1.0
